extract nav bottom row
